requirements.txt
aiogram==3.21.0
aiohttp==3.12.15
asyncpg==0.30.0
python-dotenv==1.2.2
pytest==8.3.4
pytest-asyncio==0.24.0
requests==2.32.5
